221: Rename and merge SectionCard and SectionCardListItem
