+++ lib/app/widgets/section_card.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:gruene_app/app/utils/open_url.dart';
+import 'package:gruene_app/app/utils/utils.dart';
 
-class ProfileCardListItem extends StatelessWidget {
+class SectionCardListItem extends StatelessWidget {
   final String value;
   final String? title;
   final String? url;
@@ -10,7 +11,7 @@ class ProfileCardListItem extends StatelessWidget {
   final bool copyOnTap;
   final Widget? extraTrailing;
 
-  const ProfileCardListItem({
+  const SectionCardListItem({
     super.key,
     this.title,
     required this.value,
@@ -50,6 +51,47 @@ class ProfileCardListItem extends StatelessWidget {
           extraTrailing,
           url != null ? Icon(Icons.arrow_outward, color: theme.primaryColor) : copyIcon,
         ].nonNulls.toList(),
+      ),
+    );
+  }
+}
+
+class SectionCard extends StatelessWidget {
+  final Iterable<Widget> children;
+  final String? title;
+
+  const SectionCard({super.key, required this.children, this.title});
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final title = this.title;
+
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16),
+      child: Container(
+        decoration: BoxDecoration(
+          boxShadow: [BoxShadow(color: Color.fromRGBO(0, 0, 0, 0.04), offset: Offset(0, 1), blurRadius: 12)],
+        ),
+        child: Card(
+          color: theme.colorScheme.surface,
+          elevation: 0,
+          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
+          child: Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: [
+                if (title != null)
+                  Padding(
+                    padding: const EdgeInsets.only(top: 8, left: 16, right: 16),
+                    child: Text(title, style: Theme.of(context).textTheme.titleMedium),
+                  ),
+                ...children.withDividers(Divider(indent: 16, endIndent: 16)),
+              ],
+            ),
+          ),
+        ),
       ),
     );
   }
