+++ lib/features/profiles/widgets/profile_details.dart
@@ -6,8 +6,7 @@ import 'package:gruene_app/app/utils/divisions.dart';
 import 'package:gruene_app/app/utils/open_url.dart';
 import 'package:gruene_app/app/utils/profiles.dart';
 import 'package:gruene_app/app/utils/utils.dart';
-import 'package:gruene_app/features/profiles/widgets/profile_card.dart';
-import 'package:gruene_app/features/profiles/widgets/profile_card_list_item.dart';
+import 'package:gruene_app/app/widgets/section_card.dart';
 import 'package:gruene_app/i18n/translations.g.dart';
 import 'package:gruene_app/swagger_generated_code/gruene_api.swagger.dart';
 
@@ -37,22 +36,22 @@ class ProfileDetails extends StatelessWidget {
       spacing: 16,
       children: [
         if (email != null || profile.phoneNumbers.isNotEmpty || isOwnProfile)
-          ProfileCard(
+          SectionCard(
             children: [
               if (email != null)
-                ProfileCardListItem(
+                SectionCardListItem(
                   title: t.profiles.email,
                   value: email,
                   onTap: isOwnProfile ? null : () => openMail(email, context),
                 ),
               if (profile.phoneNumbers.isNotEmpty)
-                ProfileCardListItem(title: t.profiles.phoneNumber, value: profile.phoneNumbers.first.number),
+                SectionCardListItem(title: t.profiles.phoneNumber, value: profile.phoneNumbers.first.number),
               if (isOwnProfile)
-                ProfileCardListItem(title: t.profiles.personalId, value: profile.personalId, copyOnTap: true),
+                SectionCardListItem(title: t.profiles.personalId, value: profile.personalId, copyOnTap: true),
             ],
           ),
         if (divisions.isNotEmpty)
-          ProfileCard(
+          SectionCard(
             title: t.profiles.memberships,
             children: [
               FutureLoadingScreen(
@@ -60,55 +59,58 @@ class ProfileDetails extends StatelessWidget {
                     ? () => loadDivisions(parentDivisionKeys)
                     : () async => <Division>[],
                 buildChild: (data, _) => Column(
-                  children: [...divisions, ...data].sortByLevel(reverseLevel: true).map((division) {
-                    final email = division.emails.firstOrNull?.address;
-                    return ProfileCardListItem(
-                      value: division.shortDisplayName,
-                      url: division.urls.firstOrNull,
-                      extraTrailing: email != null
-                          ? IconButton(
-                              onPressed: () => openMail(email, context),
-                              onLongPress: () => Clipboard.setData(ClipboardData(text: email)),
-                              icon: Icon(Icons.email_outlined, color: theme.primaryColor),
-                            )
-                          : null,
-                    );
-                  }).withDividers(Divider(indent: 16, endIndent: 16)),
+                  children: [...divisions, ...data]
+                      .sortByLevel(reverseLevel: true)
+                      .map((division) {
+                        final email = division.emails.firstOrNull?.address;
+                        return SectionCardListItem(
+                          value: division.shortDisplayName,
+                          url: division.urls.firstOrNull,
+                          extraTrailing: email != null
+                              ? IconButton(
+                                  onPressed: () => openMail(email, context),
+                                  onLongPress: () => Clipboard.setData(ClipboardData(text: email)),
+                                  icon: Icon(Icons.email_outlined, color: theme.primaryColor),
+                                )
+                              : null,
+                        );
+                      })
+                      .withDividers(Divider(indent: 16, endIndent: 16)),
                 ),
               ),
             ],
           ),
         if (mandateRoles.isNotEmpty)
-          ProfileCard(
+          SectionCard(
             title: t.profiles.mandateRoles,
-            children: mandateRoles.map((role) => ProfileCardListItem(value: role)),
+            children: mandateRoles.map((role) => SectionCardListItem(value: role)),
           ),
         if (officeRoles.isNotEmpty)
-          ProfileCard(
+          SectionCard(
             title: t.profiles.officeRoles,
-            children: officeRoles.map((role) => ProfileCardListItem(value: role)),
+            children: officeRoles.map((role) => SectionCardListItem(value: role)),
           ),
         if (sherpaRoles.isNotEmpty)
-          ProfileCard(
+          SectionCard(
             title: t.profiles.sherpaRoles,
-            children: sherpaRoles.map((role) => ProfileCardListItem(value: role)),
+            children: sherpaRoles.map((role) => SectionCardListItem(value: role)),
           ),
         if (profile.socialMedia.isNotEmpty)
-          ProfileCard(
+          SectionCard(
             title: t.profiles.socialMedia,
             children: profile.socialMedia.map(
-              (platform) => ProfileCardListItem(value: platform.label, url: platform.url),
+              (platform) => SectionCardListItem(value: platform.label, url: platform.url),
             ),
           ),
         if (skills.isNotEmpty)
-          ProfileCard(
+          SectionCard(
             title: t.profiles.skills,
-            children: skills.map((tag) => ProfileCardListItem(value: tag)),
+            children: skills.map((tag) => SectionCardListItem(value: tag)),
           ),
         if (interests.isNotEmpty)
-          ProfileCard(
+          SectionCard(
             title: t.profiles.interests,
-            children: interests.map((tag) => ProfileCardListItem(value: tag)),
+            children: interests.map((tag) => SectionCardListItem(value: tag)),
           ),
       ],
     );
